requirements.txt
psutil
python-ptrace
inquirer
rich
art
loguru